requirements.txt
mmcv==1.6.0
matplotlib
argparse
lpips
plyfile
pytorch_msssim
open3d
onnxruntime
onnx
onnxruntime-gpu
opencv-python
tqdm
imageio[ffmpeg]
transformers>=4.40
accelerate
